requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django==4.0
django-cors-headers==3.11.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djantic==0.7.0
drf-yasg==1.21.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.2
sqlparse==0.4.4
uritemplate==4.1.1
